Fix indexing of arrays to use appropriate types
In some platforms the "char" type may be signed, which would result in
using negative indexes into the "RESERVED" and "UNRESERVED" string
tables, resulting in incorrect expansions of URI templates e.g. when
the input data contains non-ASCII characters.

Additionally, change the loop variables to be of type "size_t", which is
the correct type to use when indexing the input strings. This change
silences GCC compiler warnings about comparing signed and unsigned
values in the loop condition, and about using a signed type as string
index.

Thanks to Pablo Saavedra <psaavedra@igalia.com> who first found the bug.